% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{clusterEnrich}
\alias{clusterEnrich}
\title{Perform Cluster Enrichment Analysis}
\usage{
clusterEnrich(
  clusterTab,
  se,
  inputSet,
  reference = NULL,
  ptm = FALSE,
  adj = "BH",
  filterP = 0.05,
  ifFDR = FALSE
)
}
\arguments{
\item{clusterTab}{A data frame containing cluster information, where each row corresponds to a gene and its assigned cluster.}

\item{se}{A SummarizedExperiment object containing gene expression data and metadata.}

\item{inputSet}{A list or data frame of gene sets to be used for enrichment analysis.}

\item{reference}{A character vector of reference genes. If NULL, it will be extracted from `se`. Default is `NULL`.}

\item{ptm}{Logical. If TRUE, the function will perform enrichment analysis on post-translational modification (PTM) gene sets. Default is `FALSE`.}

\item{adj}{Character. The method for adjusting p-values. Default is `"BH"`.}

\item{filterP}{Numeric. The p-value threshold for filtering significant results. Default is `0.05`.}

\item{ifFDR}{Logical. If TRUE, the function will use FDR-adjusted p-values for significance filtering. Default is `FALSE`.}
}
\value{
A list containing two elements:
\itemize{
  \item `table`: A data frame with enrichment results for each cluster and pathway.
  \item `plot`: A ggplot object showing the significance of enrichment for each pathway across clusters.
}
}
\description{
`clusterEnrich` performs enrichment analysis on gene clusters, using Fisher's Exact Test to determine the significance of enrichment for each cluster.
}
\details{
The function first retrieves or computes the reference set of genes or PTM sites. It then performs enrichment analysis for each cluster using the `runFisher` function.
The results are filtered based on the p-value threshold and adjusted for multiple testing if `ifFDR` is `TRUE`. The function generates a dot plot where the size and color of the points represent the significance of enrichment.
}
\examples{
# Example usage:
# Assuming `clusterTable` is a data frame with cluster assignments,
# `summarizedExp` is a SummarizedExperiment object,
# and `geneSetCollection` is a list or data frame of gene sets
# results <- clusterEnrich(clusterTable, summarizedExp, geneSetCollection, ptm = FALSE)

}
