% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{dda_example}
\alias{dda_example}
\title{dda_example}
\format{
a class S4 object of MultiAssayExperiment
}
\usage{
data(dda_example)
}
\description{
A sample of Data Dependent Acquisition (DDA) mass spectrometry data from Max Quant.
}
\examples{
data(dda_example)
}
\keyword{datasets}
