% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{dia_example}
\alias{dia_example}
\title{dia_example}
\format{
a class S4 object of MultiAssayExperiment
}
\usage{
data(dia_example)
}
\description{
A sample of Data Independent Acquisition (DIA) mass spectrometry data from
Spectronaut.
}
\examples{
data(dia_example)
}
\keyword{datasets}
