% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_FileTableQC.R
\name{fileTableQC}
\alias{fileTableQC}
\title{fileTableQC}
\usage{
fileTableQC(fileTable)
}
\arguments{
\item{fileTable}{Loaded fileTable using SmartPhos::generateInputTable
@export}
}
\description{
Check if the loaded output files (e.g. from MaxQuant), include all the columns required for the downstream assembly and include measured values (not NA or 0 for all features). This is a Quality Control (QC) check of the input files before proceeding to SmartPhos::readExperiment.
}
