% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImport.R
\name{generateInputTable}
\alias{generateInputTable}
\title{Create a template for input file list}
\usage{
generateInputTable(rawFolder)
}
\arguments{
\item{rawFolder}{a character string specifying the directory of the raw data folder}
}
\value{
This function data frame contains the input file imformation
}
\description{
This function generate a template for preparing the input file list, based on the
raw data folder structure.
}
