% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_correction.R
\name{getRatioMatrix}
\alias{getRatioMatrix}
\title{Get Ratio Matrix of Phosphoproteome Data}
\usage{
getRatioMatrix(maeData, normalization = FALSE, getAdjustedPP = FALSE)
}
\arguments{
\item{maeData}{A MultiAssayExperiment object containing phosphoproteome and full proteome data.}

\item{normalization}{A logical value indicating whether to perform normalization. Default is `FALSE`.}

\item{getAdjustedPP}{A logical value indicating whether to use adjusted phosphoproteome data. Default is `FALSE`.}
}
\value{
A numeric matrix representing the ratio of intensity of PP (phosphoproteome) data to FP (full proteome) data.
}
\description{
`getRatioMatrix` calculates the ratio matrix of phosphoproteome data from a MultiAssayExperiment object.
}
\examples{
# Example usage:
# Assuming maeData is a MultiAssayExperiment object with appropriate data
# ratio_matrix <- getRatioMatrix(maeData, normalization = TRUE, getAdjustedPP = TRUE)
# print(ratio_matrix)

}
