% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{Homo_sapien_kinase_substrate_network}
\alias{Homo_sapien_kinase_substrate_network}
\title{Homo_sapien_kinase_substrate_network}
\format{
a data.frame object
}
\usage{
data(Homo_sapien_kinase_substrate_network)
}
\value{
A \code{data frame} containing the information about the known
kinase-phosphosite interactions for Homo sapiens.
}
\description{
A prior knowledge database about the known kinase-phosphosite interactions
for Homo sapiens
}
\examples{
data(Homo_sapien_kinase_substrate_network)
}
\keyword{datasets}
