% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{Homo_sapiens}
\alias{Homo_sapiens}
\title{Homo_sapiens}
\format{
a data.frame object
}
\usage{
data(Homo_sapiens)
}
\description{
A prior knowledge database about the known kinase-phosphosite interactions for Homo sapiens
}
\examples{
data(Homo_sapiens)
}
\keyword{datasets}
