% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{Mus_musculus_kinase_substrate_network}
\alias{Mus_musculus_kinase_substrate_network}
\title{Mus_musculus_kinase_substrate_network}
\format{
a data.frame object
}
\usage{
data(Mus_musculus_kinase_substrate_network)
}
\description{
A prior knowledge database about the known kinase-phosphosite interactions for Homo sapiens
}
\examples{
data(Mus_musculus_kinase_substrate_network)
}
\keyword{datasets}
