% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImport.R
\name{normByFullProteome}
\alias{normByFullProteome}
\title{Normalise phosphoproteomic measurement by full proteome measurement}
\usage{
normByFullProteome(mae)
}
\arguments{
\item{mae}{The MultiAsssayExperiment object generated by readExperimentDIA function.}
}
\value{
This function returns a multi-assay experiment that contains the phosphoproteome and proteome data.
}
\description{
This function calculates the ratio of raw phophoproteomic intensity by the raw full proteomic intensity. The "Phosphoproteome" assay
in the MultiAssayExperiment will be replaced by the ratio.
}
