% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_visualizations.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot Heatmap of Intensity assay}
\usage{
plotHeatmap(
  type,
  se,
  data = NULL,
  top = 100,
  cutCol = 1,
  cutRow = 1,
  clustCol = TRUE,
  clustRow = TRUE,
  annotationCol,
  title
)
}
\arguments{
\item{type}{A character string specifying the type of heatmap to plot. Options are `"Top variant"`, `"Differentially expressed"`, and `"Selected time series cluster"`.}

\item{se}{A SummarizedExperiment object containing the imputed intensity assay.}

\item{data}{An optional data frame containing additional data for `"Differentially expressed"` and `"Selected time series cluster"` types. Default is `NULL`.}

\item{top}{An integer specifying the number of top variants to plot. Default is `100`.}

\item{cutCol}{An integer specifying the number of clusters for columns. Default is `1`.}

\item{cutRow}{An integer specifying the number of clusters for rows. Default is `1`.}

\item{clustCol}{A logical value indicating whether to cluster columns. Default is `TRUE`.}

\item{clustRow}{A logical value indicating whether to cluster rows. Default is `TRUE`.}

\item{annotationCol}{A character vector specifying the columns in the metadata to use for annotation. Default is `NULL`.}

\item{title}{A character string specifying the title of the heatmap.}
}
\value{
A pheatmap object showing the heatmap of Intensity data.
}
\description{
`plotHeatmap` generates a heatmap for intensity assay for different conditions, including top variants, differentially expressed genes, and selected time series clusters.
}
\details{
This function creates a heatmap using the Intensity assay from a SummarizedExperiment object. The heatmap can show the top variants based on standard deviation, differentially expressed genes, or selected time series clusters. Row normalization is performed, and the heatmap can include annotations based on specified metadata columns.
}
\examples{
# Assuming 'se' is a SummarizedExperiment object and 'data' is a data frame:
heatmap <- plotHeatmap("Top variant", se, top = 50, annotationCol = "group", title = "Top Variants Heatmap")
print(heatmap)

}
