% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_visualizations.R
\name{plotIntensity}
\alias{plotIntensity}
\title{Plot Intensity Boxplots}
\usage{
plotIntensity(se, color = "none")
}
\arguments{
\item{se}{A SummarizedExperiment object containing the assay data and metadata.}

\item{color}{A character string specifying the metadata column to use for coloring the boxplots. Default is "none".}
}
\value{
A ggplot object showing boxplots of intensities for each sample.
}
\description{
`plotIntensity` generates boxplots of assay intensities for each sample in a SummarizedExperiment object. Optionally, the boxplots can be colored based on a specified metadata column. The function handles missing values by filtering them out before plotting.
}
\examples{
# Assuming 'se' is a SummarizedExperiment object:
plot <- plotIntensity(se, color = "group")
print(plot)

}
