% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_correction.R
\name{plotLogRatio}
\alias{plotLogRatio}
\title{Plot Log Ratio of PP/FP (Phosphoproteome to Full Proteome) intensities}
\usage{
plotLogRatio(maeData, normalization = FALSE)
}
\arguments{
\item{maeData}{A MultiAssayExperiment object containing phosphoproteome and full proteome data.}

\item{normalization}{A logical value indicating whether to perform normalization. Default is `FALSE`.}
}
\value{
A ggplot object representing the boxplot of the log2 ratios.
}
\description{
`plotLogRatio` generates a boxplot of the log2 ratio of intensities of phosphoproteome to full proteome data from a MultiAssayExperiment object.
}
\examples{
# Example usage:
# Assuming maeData is a MultiAssayExperiment object with appropriate data
# plot <- plotLogRatio(maeData, normalization = TRUE)
# print(plot)

}
