% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_visualizations.R
\name{plotMissing}
\alias{plotMissing}
\title{Plot Missing Data Completeness}
\usage{
plotMissing(se)
}
\arguments{
\item{se}{A SummarizedExperiment object containing the assay data.}
}
\value{
A ggplot object showing the percentage of completeness for each sample.
}
\description{
`plotMissing` generates a bar plot showing the completeness (percentage of non-missing values) for each sample in a SummarizedExperiment object.
}
\details{
This function calculates the percentage of non-missing values for each sample in the provided SummarizedExperiment object. It then generates a bar plot where each bar represents a sample, and the height of the bar corresponds to the completeness (percentage of non-missing values) of that sample.
}
\examples{
# Assuming 'se' is a SummarizedExperiment object:
plot <- plotMissing(se)
print(plot)

}
