% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_proteome_phospho.R
\name{preprocessPhos}
\alias{preprocessPhos}
\title{Preprocess Phosphoproteome Data}
\usage{
preprocessPhos(
  seData,
  filterList = NULL,
  missCut = 50,
  transform = "log2",
  normalize = FALSE,
  getFP = FALSE,
  removeOutlier = NULL,
  assayName = NULL,
  batch = NULL,
  scaleFactorTab = NULL,
  impute = "none",
  verbose = FALSE
)
}
\arguments{
\item{seData}{A SummarizedExperiment object containing phosphoproteome data.}

\item{filterList}{A list of filters to apply on the samples. Default is `NULL`.}

\item{missCut}{Numeric value specifying the missing value cutoff percentage for filtering features. Default is `50`.}

\item{transform}{Character string specifying the transformation method ("log2", "vst", "none"). Default is `"log2"`.}

\item{normalize}{Logical value indicating whether to normalize the data. Default is `FALSE`.}

\item{getFP}{Logical value indicating whether to retrieve FP samples. Default is `FALSE`.}

\item{removeOutlier}{Character vector of samples to be removed as outliers. Default is `NULL`.}

\item{assayName}{Character string specifying the assay name in the SummarizedExperiment object. Default is `NULL`.}

\item{batch}{Character vector specifying batch effects to remove. Default is `NULL`.}

\item{scaleFactorTab}{Data frame containing scale factors for normalization. Default is `NULL`.}

\item{impute}{Character string specifying the imputation method ("QRILC", "MLE", "bpca", "missForest", "MinDet", "none"). Default is `"none"`.}

\item{verbose}{Logical value indicating whether to print detailed information. Default is `FALSE`.}
}
\value{
A SummarizedExperiment object with preprocessed phosphoproteome data.
}
\description{
`preprocessPhos` preprocesses phosphoproteome data stored in a SummarizedExperiment object by performing filtering, transformation, normalization, imputation, and batch effect removal.
}
\examples{
# Load multiAssayExperiment object
data("dia_example")
# Get SummarizedExperiment object
se <- dia_example[["Phosphoproteome"]]
colData(se) <- colData(dia_example)
# Call the function
result <- preprocessPhos(seData = se, normalize = TRUE, impute = "QRILC")

}
