% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImport.R
\name{readExperiment}
\alias{readExperiment}
\title{Read and Process the DDA experiment.}
\usage{
readExperiment(
  fileTable,
  localProbCut = 0.75,
  scoreDiffCut = 5,
  fdrCut = 0.1,
  scoreCut = 10,
  pepNumCut = 1,
  ifLFQ = TRUE,
  annotation_col = c()
)
}
\arguments{
\item{fileTable}{A data.frame containing information about the input files, including type, id, sample, and other annotations.}

\item{localProbCut}{Numeric, local probability cutoff for filtering phosphoproteomic data. Default is `0.75`.}

\item{scoreDiffCut}{Numeric, score difference cutoff for filtering phosphoproteomic data. Default is `5`.}

\item{fdrCut}{Numeric, false discovery rate cutoff for filtering proteomic data. Default is `0.1`.}

\item{scoreCut}{Numeric, score cutoff for filtering proteomic data. Default is `10`.}

\item{pepNumCut}{Numeric, peptide number cutoff for filtering proteomic data. Default is `1`.}

\item{ifLFQ}{Logical, whether to use LFQ quantification for proteomic data. Default is `TRUE`.}

\item{annotation_col}{A character vector specifying additional columns to be included in the sample annotation. Default is an empty vector.}
}
\value{
A `MultiAssayExperiment` object containing the processed phosphoproteomic and proteomic data from a DDA experiment.
}
\description{
`readExperiment` reads and processes DDA (Data-Dependent Acquisition) phosphoproteomic and proteomic data from a given file table, and returns a `MultiAssayExperiment` object.
}
\details{
The function performs the following steps:
\itemize{
  \item Reads and processes the phosphoproteomic data using the `readPhosphoExperiment` function.
  \item Reads and processes the proteomic data using the `readProteomeExperiment` function.
  \item Prepares the sample annotation table.
  \item Constructs and returns a `MultiAssayExperiment` object containing the processed data.
}
}
\examples{
# Example usage:
fileTable <- data.frame(id = c("sample1", "sample2"), sample = c("A", "B"), type = c("proteome", "proteome"))
mae <- readExperiment(fileTable)

}
