% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImport.R
\name{readExperimentDIA}
\alias{readExperimentDIA}
\title{Read and Process a DIA Experiment}
\usage{
readExperimentDIA(
  fileTable,
  localProbCut = 0.75,
  annotation_col = c(),
  onlyReviewed = TRUE,
  normalizeByProtein = FALSE
)
}
\arguments{
\item{fileTable}{A data frame containing metadata about the files to be read. Must contain columns `type`, `fileName`, `id`, and optionally `outputID`.}

\item{localProbCut}{Numeric, the local probability cutoff for phosphoproteomic data. Default is `0.75`.}

\item{annotation_col}{A character vector specifying the columns in `fileTable` to be used for sample annotation.}

\item{onlyReviewed}{A logical value indicating whether to include only reviewed proteins. Default is `TRUE`.}

\item{normalizeByProtein}{Logical, whether to normalize the data by protein. Default is `FALSE`.}
}
\value{
A `MultiAssayExperiment` object containing the processed phosphoproteome and proteome data.
}
\description{
`readExperimentDIA` reads and processes DIA (Data-Independent Acquisition) data for both phosphoproteome and proteome experiments, and constructs a `MultiAssayExperiment` object.
}
\details{
The function performs the following steps:
\itemize{
  \item Reads and processes phosphoproteomic data using `readPhosphoExperimentDIA`.
  \item Reads and processes proteomic data using `readProteomeExperimentDIA`.
  \item Prepares sample annotations based on the provided `fileTable` and `annotation_col`.
  \item Constructs a `MultiAssayExperiment` object with the processed data and sample annotations.
}
The `readPhosphoExperimentDIA` and `readProteomeExperimentDIA` functions are used to read and filter the data for phosphoproteome and proteome experiments, respectively, and they must be available in the environment.
}
\examples{
# Example usage:
fileTable <- data.frame(type = c("phosphoproteome", "proteome"), 
                        fileName = c("phos_file.txt", "prot_file.txt"), 
                        id = c("sample1", "sample2"),
                        outputID = c("s1", "s2"))
result <- readExperimentDIA(fileTable, localProbCut = 0.75, annotation_col = c("id"), normalizeByProtein = FALSE)

}
