% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{readOneProteom}
\alias{readOneProteom}
\title{Read and Process One Proteomics Sample}
\usage{
readOneProteom(inputTab, sampleName, pepNumCut = 1, ifLFQ = TRUE)
}
\arguments{
\item{inputTab}{A data.table or data.frame containing the input data for the proteomics sample.}

\item{sampleName}{A character string specifying the name of the sample to be processed.}

\item{pepNumCut}{A numeric value specifying the minimum number of peptides required for a protein to be included. Default is `1`.}

\item{ifLFQ}{A logical value indicating whether to use LFQ quantification. Default is `TRUE`.}
}
\value{
A data.table with the processed proteomics data, including columns for intensity, Uniprot ID, peptide counts, and gene names.
}
\description{
`readOneProteom` reads and processes proteomics data for a single sample, applying filters for peptide count and optionally using LFQ quantification. It returns a data.table with useful columns and unique identifiers.
}
\details{
This function processes proteomics data for a single sample by filtering based on the number of peptides and optionally using LFQ quantification. It ensures that unique identifiers are created for each protein, and removes rows with missing or zero quantification values.
}
\examples{
inputTab <- data.table::data.table(
     Intensity.sample1 = c(100, 200, NA, 50),
     LFQ.intensity.sample1 = c(90, 180, NA, 45),
     Razor...unique.peptides.sample1 = c(2, 1, 0, 3),
     Protein.IDs = c("P1", "P2", "P3", "P4"),
     Peptide.counts..all. = c(2, 1, 0, 3),
     Gene.names = c("Gene1", "Gene2", "Gene3", "Gene4")
)
result <- readOneProteom(inputTab, sampleName = "sample1", pepNumCut = 1, ifLFQ = TRUE)

}
