% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{readPhosphoExperimentDIA}
\alias{readPhosphoExperimentDIA}
\title{Read Phosphorylation Experiment Data from DIA}
\usage{
readPhosphoExperimentDIA(
  fileTable,
  localProbCut = 0.75,
  onlyReviewed = TRUE,
  showProgressBar = FALSE
)
}
\arguments{
\item{fileTable}{A data.table or data.frame containing metadata about the files to read. Must include columns `fileName`, `type`, and optionally `outputID`.}

\item{localProbCut}{A numeric value specifying the cutoff for localization probability. Default is `0.75`.}

\item{onlyReviewed}{A logical value indicating whether to include only reviewed proteins. Default is `TRUE`.}

\item{showProgressBar}{A logical value indicating whether to show a progress bar. Default is `FALSE`.}
}
\value{
A `SummarizedExperiment` object containing the processed phosphorylation data.
}
\description{
`readPhosphoExperimentDIA` reads and processes phosphorylation data from DIA experiments, applying filters for localization probability, and optionally including only reviewed proteins. It constructs a `SummarizedExperiment` object.
}
\details{
This function processes phosphorylation data from DIA experiments by filtering based on localization probability and non-zero intensity, handling multiplicity, and optionally including only reviewed proteins. The resulting data is returned as a SummarizedExperiment object with annotations and an intensity matrix.
}
\examples{
# fileTable <- data.table::data.table(fileName = c("file1.tsv", "file2.tsv"), type = "phosphoproteome")
# result <- readPhosphoExperimentDIA(fileTable, localProbCut = 0.75, onlyReviewed = TRUE, showProgressBar = TRUE)

}
