% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_correction.R
\name{runPhosphoAdjustment}
\alias{runPhosphoAdjustment}
\title{Run Phospho Adjustment}
\usage{
runPhosphoAdjustment(
  maeData,
  normalization = FALSE,
  minOverlap = 3,
  completeness = 0,
  ncore = 1
)
}
\arguments{
\item{maeData}{A MultiAssayExperiment object containing phosphoproteome and full proteome data.}

\item{normalization}{A logical value indicating whether to perform normalization. Default is `FALSE`.}

\item{minOverlap}{An integer specifying the minimum number of overlapping peptides required between samples. Default is `3`.}

\item{completeness}{A numeric value indicating the required completeness of data for features to be included. Default is `0`.}

\item{ncore}{An integer specifying the number of cores to use for parallel processing. Default is `1`.}
}
\value{
A MultiAssayExperiment object with adjusted phosphoproteome data.
}
\description{
`runPhosphoAdjustment` performs phospho adjustment on a MultiAssayExperiment object to normalize the phosphoproteome data.
}
\details{
The function performs the following steps:
\enumerate{
  \item Defines an optimization function to minimize the sum of squared differences between pairs of samples.
  \item Calculates the ratio matrix of phosphoproteome to full proteome data.
  \item Subsets features based on completeness criteria.
  \item Performs a sanity check to identify and exclude problematic samples.
  \item Sets initial values for the adjustment factor based on column medians.
  \item Estimates the adjustment factor using parallel optimization.
  \item Adjusts the phosphoproteome measurements using the estimated adjustment factor.
}
}
\examples{
# Example usage:
# Assuming maeData is a MultiAssayExperiment object with appropriate data
# adjusted_maeData <- runPhosphoAdjustment(maeData, normalization = TRUE, minOverlap = 3, completeness = 0.8, ncore = 2)
# print(adjusted_maeData)

}
