% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSmartPhos.R
\name{runSmartPhos}
\alias{runSmartPhos}
\title{Launch the SmartPhos Shiny Application}
\usage{
runSmartPhos()
}
\value{
The function does not return a value; it starts the Shiny application for SmartPhos.
}
\description{
`runSmartPhos` launches the SmartPhos Shiny application, which provides an interactive interface for analyzing phosphoproteomic data.
}
\details{
The `runSmartPhos` function locates the Shiny app directory within the `SmartPhos` package and launches the application. 
If the app directory cannot be found, the function will stop and prompt the user to reinstall the `SmartPhos` package.
}
\examples{
# To run the SmartPhos Shiny application, simply call:
runSmartPhos()
}
