% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{swissProt}
\alias{swissProt}
\title{swissProt}
\format{
an object of "tbl_df" (tidy table)
}
\usage{
data(swissProt)
}
\value{
A \code{data frame} or \code{tibble} containing high-level
annotations for manually curated proteins.
}
\description{
This is a high-quality, manually curated protein sequence database which
provides a high level of annotations (such as the description of the function
of a protein, structure of its domains, post-translational modifications,
variants, etc.), a minimal level of redundancy and high level of integration
with other databases.
}
\examples{
data(swissProt)
}
\keyword{datasets}
