% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_correction.R
\name{checkRatioMat}
\alias{checkRatioMat}
\title{Check the PP/FP ratio matrix and remove feature that do not meet requirements}
\usage{
checkRatioMat(ratioMat, minOverlap = 3)
}
\arguments{
\item{ratioMat}{A numeric matrix representing the ratio of phosphoproteome data to full proteome data.}

\item{minOverlap}{An integer specifying the minimum number of overlapping peptides required between samples. Default is `3`.}
}
\value{
A character vector of sample names that do not meet the overlap criteria.
}
\description{
`checkRatioMat` checks the ratio matrix for samples that do not have sufficient overlap of phospho-peptides between enriched (PP) and unenriched (FP) samples.
}
\examples{
# Example usage:
# Assuming ratioMat is a matrix with appropriate data
# excluded_samples <- checkRatioMat(ratioMat, minOverlap = 3)
# print(excluded_samples)

}
