% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinase_activity.R
\name{getDecouplerNetwork}
\alias{getDecouplerNetwork}
\title{Load Kinase-Substrate Interaction Network}
\usage{
getDecouplerNetwork(speciesRef)
}
\arguments{
\item{speciesRef}{A character string specifying the species. Supported values are "Homo sapiens" and "Mus musculus".}
}
\value{
A data frame containing the kinase-substrate interaction network for the specified species.
}
\description{
`getDecouplerNetwork` loads the kinase-substrate interaction network for a specified species from pre-defined files.
}
\examples{
# Load the human kinase-substrate interaction network
human_network <- getDecouplerNetwork("Homo sapiens")

# Load the mouse kinase-substrate interaction network
mouse_network <- getDecouplerNetwork("Mus musculus")

}
