% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_clustering.R
\name{mscale}
\alias{mscale}
\title{Scale and Center a Matrix}
\usage{
mscale(x, center = TRUE, scale = TRUE, censor = NULL, useMad = FALSE)
}
\arguments{
\item{x}{A numeric matrix where rows are features and columns are samples.}

\item{center}{Logical. If TRUE, the rows are centered by subtracting the mean or median. Default is `TRUE`.}

\item{scale}{Logical. If TRUE, the rows are scaled by dividing by the standard deviation or mean absolute deviation. Default is `TRUE`.}

\item{censor}{A numeric vector of length one or two for censoring the scaled values. 
If length one, values are censored symmetrically at positive and negative values. 
If length two, the first value is the lower limit and the second value is the upper limit. Default is `NULL`.}

\item{useMad}{Logical. If TRUE, the mean absolute deviation is used for scaling instead of the standard deviation. Default is `FALSE`.}
}
\value{
A scaled and centered numeric matrix with the same dimensions as the input matrix `x`.
}
\description{
`mscale` scales and centers each row of a matrix, with options for using mean or median, standard deviation or mean absolute deviation, and censoring extreme values.
}
\details{
The function allows for flexible scaling and centering of the rows of a matrix:
\itemize{
  \item If both `center` and `scale` are TRUE, rows are centered and scaled.
  \item If only `center` is TRUE, rows are centered but not scaled.
  \item If only `scale` is TRUE, rows are scaled but not centered.
  \item If neither `center` nor `scale` is TRUE, the original matrix is returned.
}
The function can also censor extreme values, either symmetrically or asymmetrically, based on the `censor` parameter.
}
\examples{
# Example usage:
# Assuming `dataMatrix` is a numeric matrix with expression data
# scaledMatrix <- mscale(dataMatrix, center = TRUE, scale = TRUE, censor = 2)

}
