% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_clustering.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Plot Time Series Data for a gene or phospho site from SummarizedExperiment object}
\usage{
plotTimeSeries(
  se,
  type,
  geneID,
  symbol,
  condition,
  treatment,
  refTreat,
  addZero = FALSE,
  zeroTreat = NULL,
  timerange
)
}
\arguments{
\item{se}{A SummarizedExperiment object containing the data.}

\item{type}{The type of plot to generate. Options are `"expression"`, `"logFC"`, or `"two-condition expression"`.}

\item{geneID}{The identifier of the gene or feature to plot.}

\item{symbol}{The symbol or name of the gene or feature to use as the plot title.}

\item{condition}{The condition corresponds to one of the columns from the colData of SE object.}

\item{treatment}{The treatment to use for filtering the data.}

\item{refTreat}{The reference treatment to compare against.}

\item{addZero}{Logical, whether to add a zero time point to the data. Default is `FALSE`.}

\item{zeroTreat}{The treatment to use for adding the zero time point. Default is `NULL`.}

\item{timerange}{The range of time points to include in the plot.}
}
\value{
A ggplot object representing the time series plot.
}
\description{
`plotTimeSeries` plots time series data for a given gene or phospho site from a given SummarizedExperiment object, allowing different types of plots such as expression, log fold change, or two-condition expression.
}
\details{
This function generates time series plots for a specified gene or feature from a SummarizedExperiment (SE) object. The type of plot can be one of the following:
- "expression": Plots normalized expression levels over time.
- "logFC": Plots log fold change (logFC) over time, comparing a treatment to a reference treatment.
- "two-condition expression": Plots normalized expression levels over time for two conditions.

The function can add a zero time point if specified and handles data with and without subject-specific information. The plot includes points for each time point and a summary line representing the mean value.

The x-axis represents time, and the y-axis represents the selected metric (normalized expression or logFC). The plot is customized with various aesthetic elements, such as point size, line type, axis labels, and title formatting.
}
\examples{
# Assuming 'se' is a SummarizedExperiment object with appropriate data:
plotTimeSeries(se, type = "expression", geneID = "gene1", symbol = "Gene 1", condition = "treatment", treatment = "A", refTreat = "B", timerange = c("0h", "1h", "2h"))

}
