% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_proteome_phospho.R
\name{preprocessProteome}
\alias{preprocessProteome}
\title{Preprocess Proteome Data}
\usage{
preprocessProteome(
  seData,
  filterList = NULL,
  missCut = 50,
  transform = "log2",
  normalize = FALSE,
  getPP = FALSE,
  removeOutlier = NULL,
  impute = "QRILC",
  batch = NULL,
  verbose = FALSE,
  scaleFactorTab = NULL
)
}
\arguments{
\item{seData}{A SummarizedExperiment object containing proteome data.}

\item{filterList}{A list of filters to apply on the samples. Default is `NULL`.}

\item{missCut}{Numeric value specifying the missing value cutoff percentage for filtering features. Default is `50`.}

\item{transform}{Character string specifying the transformation method ("log2", "vst", "none"). Default is `"log2"`.}

\item{normalize}{Logical value indicating whether to normalize the data. Default is `FALSE`.}

\item{getPP}{Logical value indicating whether to retrieve PP samples. Default is `FALSE`.}

\item{removeOutlier}{Character vector of samples to be removed as outliers. Default is `NULL`.}

\item{impute}{Character string specifying the imputation method ("QRILC", "MLE", "bpca", "missForest", "MinDet", "none"). Default is `"QRILC"`.}

\item{batch}{Character vector specifying batch effects to remove. Default is `NULL`.}

\item{verbose}{Logical value indicating whether to print detailed information. Default is `FALSE`.}

\item{scaleFactorTab}{Data frame containing scale factors for normalization. Default is `NULL`.}
}
\value{
A SummarizedExperiment object with preprocessed proteome data.
}
\description{
`preprocessProteome` preprocesses proteome data stored in a SummarizedExperiment object by performing filtering, transformation, normalization, imputation, and batch effect removal.
}
\examples{
# Example usage:
# Assuming seData is a SummarizedExperiment object with appropriate data
# processedData <- preprocessProteome(seData, filterList = list(sampleType = "FP"), missCut = 30, normalize = TRUE)

}
