% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImport.R
\name{readExperiment}
\alias{readExperiment}
\title{Read the whole experiment}
\usage{
readExperiment(
  fileTable,
  localProbCut = 0.75,
  scoreDiffCut = 5,
  fdrCut = 0.1,
  scoreCut = 10,
  pepNumCut = 1,
  ifLFQ = TRUE,
  annotation_col = c()
)
}
\arguments{
\item{fileTable}{a data frame that contains the information of raw data files.
This table is normally generated by \code{generateInputTable} function.}

\item{localProbCut}{a numeric variable, the cut-off for localization probability for phosphoproteome data.
The default value is 0.75.}

\item{scoreDiffCut}{a numeric variable, the cut-off of score difference for phosphoproteome data.
The default value is 5.}

\item{fdrCut}{a numeric variable, the FDR cut-off for proteome data. The default value is 0.1.}

\item{scoreCut}{a numeric variable, the cut-off of protein scoring for proteome data. The default value is 10.}

\item{pepNumCut}{a numeric variable, the minial number of peptides required for quantifying a protein. The default value is 1.}

\item{ifLFQ}{a logic variable, whether to use LFQ normalization from MaxQuant. Default is TRUE.}

\item{annotation_col}{a charachter vector, the names of columns in the input file table that can be
read in as sample annotations.}
}
\value{
This function returns a multi-assay experiment that contains the phosphoproteome and proteome data.
}
\description{
Function to read the whole experiment and create multi-assay experiment object
}
