% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataImport.R
\name{readExperimentDIA}
\alias{readExperimentDIA}
\title{Read the whole experiment (DIA)}
\usage{
readExperimentDIA(
  fileTable,
  localProbCut = 0.75,
  annotation_col = c(),
  normalizeByProtein = FALSE
)
}
\arguments{
\item{fileTable}{a data frame that contains the information of raw data files.
This table is normally generated by \code{generateInputTable} function.}

\item{localProbCut}{a numeric variable, the cut-off for localization probability for phosphoproteome data.
The default value is 0.75.}

\item{annotation_col}{a character vector, the names of columns in the input file table that can be
read in as sample annotations.}
}
\value{
This function returns a multi-assay experiment that contains the phosphoproteome and proteome data.
}
\description{
Function to read the whole experiment and create multi-assay experiment object
}
