% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{readProteomeExperiment}
\alias{readProteomeExperiment}
\title{Read and Process Proteomics Experiment Data}
\usage{
readProteomeExperiment(
  fileTable,
  fdrCut = 0.1,
  scoreCut = 10,
  pepNumCut = 1,
  ifLFQ = TRUE
)
}
\arguments{
\item{fileTable}{A data.table or data.frame containing the file information with columns for file names, sample names, and IDs.}

\item{fdrCut}{A numeric value specifying the maximum false discovery rate (FDR) threshold. Default is `0.1`.}

\item{scoreCut}{A numeric value specifying the minimum score threshold. Default is `10`.}

\item{pepNumCut}{A numeric value specifying the minimum number of peptides required for a protein to be included. Default is `1`.}

\item{ifLFQ}{A logical value indicating whether to use LFQ quantification. Default is `TRUE`.}
}
\value{
A `SummarizedExperiment` object containing the processed proteomics data.
}
\description{
`readProteomeExperiment` reads and processes proteomics data from multiple samples, applying various quality filters, and returns a `SummarizedExperiment` object.
}
\details{
This function processes proteomics data by filtering based on FDR, score, and peptide count, and optionally using LFQ quantification. It aggregates the data from multiple samples and constructs a SummarizedExperiment object.
}
\examples{
# se <- readProteomeExperiment(fileTable)

}
