% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{runFisher}
\alias{runFisher}
\title{Perform Fisher's Exact Test on Gene Sets}
\usage{
runFisher(genes, reference, inputSet, ptm = FALSE)
}
\arguments{
\item{genes}{A character vector of genes of interest.}

\item{reference}{A character vector of reference genes.}

\item{inputSet}{A list containing gene set collections. If `ptm` is TRUE, this should be a data frame with specific columns.}

\item{ptm}{Logical. If TRUE, perform the test on post-translational modification (PTM) gene sets. Default is `FALSE`.}
}
\value{
A data frame with the results of the Fisher's Exact Test, including the gene set name, the number of genes in the set, set size, p-value, adjusted p-value, and the genes in the set.
}
\description{
`runFisher` performs Fisher's Exact Test to determine the enrichment of a set of genes within reference gene sets.
}
\details{
The function can operate in two modes: standard gene sets and PTM-specific gene sets. For PTM-specific gene sets, additional filtering and processing are performed.
}
\examples{
# Example usage:
# Assuming `genesOfInterest` is a vector of genes and `referenceGenes` is a vector of reference genes
# and `geneSetCollection` is a list of gene sets
# results <- runFisher(genesOfInterest, referenceGenes, geneSetCollection, ptm = FALSE)

}
