% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{runSmartPhos}
\alias{runSmartPhos}
\title{Launch the SmartPhos Shiny Application}
\usage{
runSmartPhos()
}
\value{
The function does not return a value; it starts the Shiny application for \code{SmartPhos}.
}
\description{
\code{runSmartPhos} launches the SmartPhos \code{Shiny} application, which provides an interactive interface for analyzing phosphoproteomic data.
}
\details{
The \code{runSmartPhos} function locates the \code{Shiny} app directory within the \code{SmartPhos} package and launches the application.
If the app directory cannot be found, the function will stop and prompt the user to re-install the \code{SmartPhos} package.
}
\examples{
# To run the SmartPhos Shiny application, simply call:
# runSmartPhos()
}
